requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.17.0
openai==1.3.0